Fixed registration
Registration had many errors.
Users can now register on the site, and a profile is automatically created.

Changed the login redirect url to go to profile page.
After registration users are directed to the login page as they are not automatically logged in. This is to ensure that they remember their credentials, and that if we implement email verification they cannot access their account yet.