Added connecting game accounts
Game accounts can now be connected, still need to implement displaying them.
Had to clear migrations to do this.
Added a check in profile to ensure that a profile exists when trying to view it.
Changed login to also provide a link to signup.
Added filtering to settings.py so we can filter API results based on parameters.