Added received_ratings to profile, updated ratings form, updated algorithm
Profile now has a received_ratings column, used by the algorithm instead of the sessions played.
This addition was made due to the issue of the algorithm dividing commendations by sessions played, as (should there be more than 2 people in a session) the algorithm would not give a valid result.
The addition of the received_ratings column fixes this, as we now increment it after every player's match rating, so we have a total amount of times that they've been rated.